+++ Content/Python/UEFN_Toolbelt/tools/__init__.py
@@ -58,3 +58,5 @@ from . import actor_org_tools
 from . import advanced_alignment
 from . import zone_tools
 from . import proximity_tools
+from . import postprocess_tools
+from . import audio_tools